[수정] 이미지 다운로드 확장 수정
