+++ SOOUM/SOOUM/Presentations/Main/Profile/UpdateProfile/UpdateProfileViewController.swift
@@ -7,6 +7,7 @@
 
 import UIKit
 
+import Kingfisher
 import SnapKit
 import Then
 import YPImagePicker
@@ -100,7 +101,7 @@ class UpdateProfileViewController: BaseNavigationViewController, View {
     
     func bind(reactor: UpdateProfileViewReactor) {
         
-        UIImage().download(strUrl: reactor.profile.profileImg?.url) { [weak self] image in
+        KingfisherManager.shared.download(strUrl: reactor.profile.profileImg?.url) { [weak self] image in
             self?.updateProfileView.image = image
         }
         self.updateProfileView.text = reactor.profile.nickname
